feat: 로그인 클래스 리팩터링
